% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcta_data_functions.R
\name{add_aadt_data}
\alias{add_aadt_data}
\title{Get Traffic Data}
\usage{
add_aadt_data(data)
}
\arguments{
\item{data}{data.frame or tibble with column called 'zcta' at minimum.}
}
\value{
the input data.frame with the following columns appended

\code{moving_roads_density} density of roads with moving traffic
(meters of road per square meter of area)

\code{stop_go_roads_density} density of roads with stop and go
traffic (meters of road per square meter of area)

\code{moving_traffic_density} moving traffic density (vehicle-meters
per square meter of area)

\code{stop_go_traffic_density} stop and go traffic density
(vehicle-meters per square meter of area)

\code{moving_truck_density} moving truck traffic density
(truck-meters per square meter of area)

\code{stop_go_truck_density} stop and go truck traffic
density (truck-meters per square meter of area)
}
\description{
adds Average Annual Daily Traffic Density. AADT data
is from 2017 and corresponds to 2010 ZCTA vintages.
}
\examples{
my_data <- data.frame(zcta = c('45229', '45056', '47012'))
add_aadt_data(data = my_data)
}
\references{
https://degauss.org/aadt
}
