% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcta_data_functions.R
\name{add_depindex_data}
\alias{add_depindex_data}
\title{Get Community Deprivation Data}
\usage{
add_depindex_data(data)
}
\arguments{
\item{data}{data.frame or tibble with column called 'zcta' at minimum.
Optionally, another column called 'year', with possible values 2000 to
2019. If no year is given, defaults to 2018}
}
\value{
the input data.frame with the following columns appended

\code{zcta_year} the ZCTA boundary year matched to the input year (2000 or 2010)

\code{dep_year} the ACS data year matched to the input year (2015 or 2018)

\code{fraction_assisted_income} fraction of households receiving public assistance income
or food stamps or SNAP in the past 12 months

\code{fraction_high_school_edu} fraction of population 25 and older with educational attainment
of at least high school graduation (includes GED equivalency)

\code{median_income} median household income in the past 12 months in 2015 inflation-adjusted dollars

\code{fraction_no_health_ins} fraction of poulation with no health insurance coverage

\code{fraction_poverty} fraction of population with income in past 12 months below poverty level

\code{fraction_vacant_housing} fraction of houses that are vacant

\code{dep_index} a composite measure of the other 6 variables
}
\description{
\code{add_depindex_data()} adds community deprivation index and
related variables from the American Community Survey (ACS). Data is
available for 2000 and 2010 ZCTAs and for years 2015 and 2018.
}
\examples{
my_data <- data.frame(zcta = c('45229', '45056', '47012'))
add_depindex_data(data = my_data)
}
\references{
https://geomarker.io/dep_index/
}
