% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcta_data_functions.R
\name{add_ejscreen_data}
\alias{add_ejscreen_data}
\title{Get EJ Screen Data}
\usage{
add_ejscreen_data(data)
}
\arguments{
\item{data}{data.frame or tibble with column called 'zcta' at minimum.
Optionally, another column called 'year', with possible values 2000 to
2019. If no year is given, defaults to 2018.}
}
\description{
\code{add_ejscreen_data()} adds traffic proximity index, ozone concentration,
particulate matter (PM) concentration, PM from diesel traffic, and a
respiratory hazard index from the EPA EJ Screen database. Data is available for
2000 and 2010 ZCTAs and for years 2015 to 2018.
}
\examples{
my_data <- data.frame(zcta = c('45229', '45056', '47012'))
add_ejscreen_data(data = my_data)
}
\references{
https://www.epa.gov/ejscreen/overview-environmental-indicators-ejscreen
}
