% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcta_data_functions.R
\name{add_landcover_data}
\alias{add_landcover_data}
\title{Get Land Cover Data}
\usage{
add_landcover_data(data)
}
\arguments{
\item{data}{data.frame or tibble with column called 'zcta' at minimum.
Optionally, another column called 'year', with possible values 2000 to
2019. If no year is given, defaults to 2016.}
}
\value{
the input data.frame with the following columns appended

\code{zcta_year} the ZCTA boundary year matched to the input year (2000 or 2010)

\code{nlcd_year} the National Landcover data year matched to the input year (2001, 2006, 2011, or 2016)

\code{treecanopy_year} if tree canopy data was not available in the nlcd_year, the
year of the tree canopy data, otherwise NA

\code{pct_green} percentage of pixels within ZCTA classified as green (defined as all NLCD land cover
codes except water, ice/snow, developed medium intensity, developed high intensity, and rock/sand/clay)

\code{pct_impervious} average percent impervious for pixels within ZCTA

\code{pct_treecanopy} average percent tree canopy for pixels within ZCTA
}
\description{
\code{add_landcover_data()} adds percentage of land classified as 'green',
percentage impervious land, and percentage of land covered by tree canopy
from the National Landcover Database. Data is available for
2000 and 2010 ZCTAs and for years 2001, 2006, 2011, and 2016.
}
\examples{
my_data <- data.frame(zcta = c('45229', '45056', '47012'))
add_landcover_data(data = my_data)
}
\references{
https://www.mrlc.gov/
}
