% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcta_data_functions.R
\name{add_narr_cell_zcta}
\alias{add_narr_cell_zcta}
\title{Get NARR Cell ID for ZCTA Centroid}
\usage{
add_narr_cell_zcta(data)
}
\arguments{
\item{data}{data.frame or tibble with column called 'zcta' at minimum.}
}
\value{
the input data.frame with the following columns appended

\code{narr_cell} NARR cell indentifier corresponding to the
NARR cell that intersects the ZCTA centroid
}
\description{
adds NARR cell identifier based on ZCTA centroid for subsequent
use with the {addNarrData} package
}
\examples{
my_data <- data.frame(zcta = c('45229', '45056', '47012'))
add_narr_cell_zcta(data = my_data)
}
\references{
https://geomarker.io/addNarrData/
}
