% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcta_data_functions.R
\name{add_race_data}
\alias{add_race_data}
\title{Get Racial Composition Data}
\usage{
add_race_data(data)
}
\arguments{
\item{data}{data.frame or tibble with column called 'zcta' at minimum.}
}
\value{
the input data.frame with the following columns appended

\code{total} total population of ZCTA

\code{wnh} number of white non-Hispanic residents

\code{bnh} number of black non-Hispanic residents

\code{racial_ice} racial index of concentration at the extremes \code{(wnh-bnh)/total)}

\code{pct_wnh} percent white \code{wnh/totalx100}

\code{pct_bnh} percent black \code{bnh/totalx100}
}
\description{
adds numbers and percentages of white non-Hispanic
and black non-Hispanic residents from the 2018 ACS for given ZIP codes,
as well as a racial index of concentration at the extremes (ICE). Data is
available for 2010 ZCTAs in the contiguous United States.
}
\examples{
my_data <- data.frame(zcta = c('45229', '45056', '47012'))
add_race_data(data = my_data)
}
