% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcta_data_functions.R
\name{add_road_data}
\alias{add_road_data}
\title{Get Length and Density of Roads}
\usage{
add_road_data(data)
}
\arguments{
\item{data}{data.frame or tibble with column called 'zcta' at minimum.
Optionally, another column called 'year', with possible values 2000 to
2019. Years 2000 to 2009 will be matched to 2000 ZCTAs and years 2010 to
2019 will be matched to 2010 ZCTAs. If no year is given, defaults to 2018.}
}
\value{
the input data.frame with the following columns appended

\code{primary_road_length} length (m) of primary roads in ZCTA

\code{primary_road_density} length(m) of primary roads in ZCTA divided by area of ZCTA (km2)

\code{secondary_road_length} length (m) of secondary roads in ZCTA

\code{secondary_road_density} length(m) of secondary roads in ZCTA divided by area of ZCTA (km2)
}
\description{
\code{add_road_data()} adds length (m) and density (m/km2) of primary and
secondary roads as defined by 2018 TIGER Line files. Data is available for
2000 and 2010 ZCTAs.
}
\examples{
my_data <- data.frame(zcta = c('45229', '45056', '47012'))
add_road_data(data = my_data)
}
